requirements.txt
absl-py==2.1.0
accelerate==1.0.1
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2
astunparse==1.6.3
attrs==24.2.0
bitsandbytes==0.44.1
certifi==2024.8.30
charset-normalizer==3.4.0
contourpy==1.3.1
cycler==0.12.1
dbus-python==1.3.2
distro==1.9.0
einops==0.8.0
fastdtw==0.3.4
filelock==3.16.1
flatbuffers==24.3.25
fonttools==4.55.3
frozenlist==1.4.1
fsspec==2024.9.0
gast==0.6.0
google-pasta==0.2.0
greenlet==3.1.1
grpcio==1.66.2
h11==0.14.0
h5py==3.12.1
httpcore==1.0.6
httpx==0.27.2
huggingface-hub==0.25.2
idna==3.10
Jinja2==3.1.4
jiter==0.6.1
jsonpatch==1.33
jsonpointer==3.0.0
keras==3.6.0
kiwisolver==1.4.7
langchain==0.3.3
langchain-core==0.3.10
langchain-openai==0.2.2
langchain-text-splitters==0.3.0
langsmith==0.1.134
libclang==18.1.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.1
matplotlib==3.10.0
mdurl==0.1.2
ml-dtypes==0.4.1
mpmath==1.3.0
multidict==6.1.0
namex==0.0.8
networkx==3.4.1
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.77
nvidia-nvtx-cu12==12.1.105
openai==1.51.2
opt_einsum==3.4.0
optree==0.13.0
orjson==3.10.7
packaging==24.1
pandas==2.2.3
pillow==10.4.0
propcache==0.2.0
protobuf==4.25.5
psutil==6.0.0
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
PyGObject==3.48.2
pyparsing==3.2.0
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
regex==2024.9.11
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.2
safetensors==0.4.5
scipy==1.14.1
setuptools==68.1.2
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.35
sympy==1.13.3
tabulate==0.9.0
tenacity==8.5.0
tensorboard==2.17.1
tensorboard-data-server==0.7.2
tensorflow==2.17.0
termcolor==2.5.0
tiktoken==0.8.0
tokenizers==0.20.1
torch==2.4.1
torchvision==0.19.1
tqdm==4.66.5
transformers==4.45.2
triton==3.0.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
Werkzeug==3.0.4
wheel==0.42.0
wrapt==1.16.0
yarl==1.15.2
